refactor(ui): extract Drawer from Modal for resizable side panels
Separates resizable slideover functionality into a dedicated Drawer
component, simplifying Modal back to a pure dialog overlay. Migrates
TracePage, SessionPage, and EvaluatorTracePage to use Drawer.
