+++ app/src/components/core/overlay/constants.ts
@@ -1,8 +1,8 @@
 import type { SizeValue } from "@phoenix/types/sizing";
 
 /**
- * Minimum size for fullscreen-style resizable slideovers (e.g. trace,
- * session, evaluator detail drawers). Import this constant so every
- * consumer stays in sync. Expressed as a percentage of the viewport width.
+ * Minimum size for resizable drawers (e.g. trace, session, evaluator
+ * detail drawers). Import this constant so every consumer stays in sync.
+ * Expressed as a percentage of the viewport width.
  */
-export const SLIDEOVER_MIN_SIZE: SizeValue = "40%";
+export const DRAWER_MIN_SIZE: SizeValue = "40%";
