+++ app/src/components/core/overlay/index.tsx
@@ -1,6 +1,7 @@
 export * from "./PopoverArrow";
 export * from "./Popover";
 export * from "./Modal";
-export * from "./ModalContext";
+export * from "./Drawer";
+export * from "./DrawerContext";
 export * from "./constants";
-export * from "./useDefaultModalSize";
+export * from "./useDefaultDrawerSize";
