+++ app/src/pages/dataset/evaluators/EvaluatorTracePage.tsx
@@ -4,10 +4,9 @@ import invariant from "tiny-invariant";
 
 import {
   Dialog,
+  Drawer,
   Flex,
   Loading,
-  Modal,
-  ModalOverlay,
 } from "@phoenix/components";
 import {
   DialogCloseButton,
@@ -16,8 +15,8 @@ import {
   DialogTitle,
   DialogTitleExtra,
 } from "@phoenix/components/core/dialog";
-import { SLIDEOVER_MIN_SIZE } from "@phoenix/components/core/overlay/constants";
-import { useDefaultModalSize } from "@phoenix/components/core/overlay/useDefaultModalSize";
+import { DRAWER_MIN_SIZE } from "@phoenix/components/core/overlay/constants";
+import { useDefaultDrawerSize } from "@phoenix/components/core/overlay/useDefaultDrawerSize";
 import { ShareLinkButton } from "@phoenix/components/ShareLinkButton";
 import { TraceDetails } from "@phoenix/pages/trace/TraceDetails";
 
@@ -35,7 +34,7 @@ export function EvaluatorTracePage() {
     EVALUATOR_DETAILS_ROUTE_ID
   );
   const projectId = loaderData?.projectId;
-  const { defaultSize, onSizeChange } = useDefaultModalSize({
+  const { defaultSize, onSizeChange } = useDefaultDrawerSize({
     id: "evaluator-trace-details",
   });
 
@@ -45,45 +44,38 @@ export function EvaluatorTracePage() {
   invariant(evaluatorId, "evaluatorId is required");
 
   return (
-    <ModalOverlay
+    <Drawer
       isOpen
-      onOpenChange={(isOpen) => {
-        if (!isOpen) {
-          navigate(`/datasets/${datasetId}/evaluators/${evaluatorId}`);
-        }
-      }}
+      onClose={() =>
+        navigate(`/datasets/${datasetId}/evaluators/${evaluatorId}`)
+      }
+      defaultSize={defaultSize}
+      minSize={DRAWER_MIN_SIZE}
+      onResize={onSizeChange}
     >
-      <Modal
-        variant="slideover"
-        isResizable
-        defaultSize={defaultSize}
-        minSize={SLIDEOVER_MIN_SIZE}
-        onResize={onSizeChange}
-      >
-        <Dialog>
-          {({ close }) => (
-            <DialogContent>
-              <DialogHeader>
-                <Flex direction="row" gap="size-200" alignItems="center">
-                  <DialogCloseButton close={close} />
-                  <DialogTitle>Trace Details</DialogTitle>
-                </Flex>
-                <DialogTitleExtra>
-                  <ShareLinkButton
-                    preserveSearchParams
-                    buttonText="Share"
-                    tooltipText="Copy trace link to clipboard"
-                    successText="Trace link copied to clipboard"
-                  />
-                </DialogTitleExtra>
-              </DialogHeader>
-              <Suspense fallback={<Loading />}>
-                <TraceDetails traceId={traceId} projectId={projectId} />
-              </Suspense>
-            </DialogContent>
-          )}
-        </Dialog>
-      </Modal>
-    </ModalOverlay>
+      <Dialog>
+        {({ close }) => (
+          <DialogContent>
+            <DialogHeader>
+              <Flex direction="row" gap="size-200" alignItems="center">
+                <DialogCloseButton close={close} />
+                <DialogTitle>Trace Details</DialogTitle>
+              </Flex>
+              <DialogTitleExtra>
+                <ShareLinkButton
+                  preserveSearchParams
+                  buttonText="Share"
+                  tooltipText="Copy trace link to clipboard"
+                  successText="Trace link copied to clipboard"
+                />
+              </DialogTitleExtra>
+            </DialogHeader>
+            <Suspense fallback={<Loading />}>
+              <TraceDetails traceId={traceId} projectId={projectId} />
+            </Suspense>
+          </DialogContent>
+        )}
+      </Dialog>
+    </Drawer>
   );
 }
