+++ app/src/pages/trace/TracePage.tsx
@@ -4,10 +4,9 @@ import { useNavigate, useParams, useSearchParams } from "react-router";
 import {
   Button,
   Dialog,
+  Drawer,
   Flex,
   Loading,
-  Modal,
-  ModalOverlay,
   TitleWithID,
 } from "@phoenix/components";
 import { PxiGlyph } from "@phoenix/components/agent/PxiGlyph";
@@ -18,8 +17,8 @@ import {
   DialogTitle,
   DialogTitleExtra,
 } from "@phoenix/components/core/dialog";
-import { SLIDEOVER_MIN_SIZE } from "@phoenix/components/core/overlay/constants";
-import { useDefaultModalSize } from "@phoenix/components/core/overlay/useDefaultModalSize";
+import { DRAWER_MIN_SIZE } from "@phoenix/components/core/overlay/constants";
+import { useDefaultDrawerSize } from "@phoenix/components/core/overlay/useDefaultDrawerSize";
 import { ShareLinkButton } from "@phoenix/components/ShareLinkButton";
 import { SELECTED_SPAN_NODE_ID_PARAM } from "@phoenix/constants/searchParams";
 import { useAgentContext } from "@phoenix/contexts/AgentContext";
@@ -40,7 +39,7 @@ export function TracePage() {
   const setIsOpen = useAgentContext((state) => state.setIsOpen);
   const { rootPath, tab } = useProjectRootPath();
   const selectedSpanNodeId = searchParams.get(SELECTED_SPAN_NODE_ID_PARAM);
-  const { defaultSize, onSizeChange } = useDefaultModalSize({
+  const { defaultSize, onSizeChange } = useDefaultDrawerSize({
     id: "trace-details",
   });
 
@@ -49,63 +48,54 @@ export function TracePage() {
   const paginationSubjectId = selectedSpanNodeId || traceId;
 
   return (
-    <ModalOverlay
+    <Drawer
       isOpen
-      onOpenChange={(isOpen) => {
-        if (!isOpen) {
-          navigate(`${rootPath}/${tab}`);
-        }
-      }}
+      onClose={() => navigate(`${rootPath}/${tab}`)}
+      defaultSize={defaultSize}
+      minSize={DRAWER_MIN_SIZE}
+      onResize={onSizeChange}
     >
-      <Modal
-        variant="slideover"
-        isResizable
-        defaultSize={defaultSize}
-        minSize={SLIDEOVER_MIN_SIZE}
-        onResize={onSizeChange}
-      >
-        <Dialog>
-          {({ close }) => (
-            <DialogContent>
-              <DialogHeader>
-                <Flex direction="row" gap="size-200" alignItems="center">
-                  <DialogCloseButton close={close} />
-                  <TraceDetailsPaginator currentId={paginationSubjectId} />
-                  <DialogTitle>
-                    <TitleWithID title="Trace" id={traceId as string} />
-                  </DialogTitle>
-                </Flex>
-                <DialogTitleExtra>
-                  {isAgentsEnabled ? (
-                    /* The global FAB is intentionally hidden while a modal overlay
-                        is open, so traces expose a local PXI entrypoint here. */
-                    <Button
-                      size="S"
-                      variant="primary"
-                      leadingVisual={<PxiGlyph variant="resting" />}
-                      onPress={() => setIsOpen(true)}
-                    >
-                      Ask PXI
-                    </Button>
-                  ) : null}
-                  <ShareLinkButton
-                    preserveSearchParams
-                    buttonText="Share"
-                    tooltipText="Copy trace link to clipboard"
-                    successText="Trace link copied to clipboard"
-                  />
-                </DialogTitleExtra>
-              </DialogHeader>
-              <Suspense fallback={<Loading />}>
-                <TraceDetails
-                  traceId={traceId as string}
-                  projectId={projectId as string}
+      <Dialog>
+        {({ close }) => (
+          <DialogContent>
+            <DialogHeader>
+              <Flex direction="row" gap="size-200" alignItems="center">
+                <DialogCloseButton close={close} />
+                <TraceDetailsPaginator currentId={paginationSubjectId} />
+                <DialogTitle>
+                  <TitleWithID title="Trace" id={traceId as string} />
+                </DialogTitle>
+              </Flex>
+              <DialogTitleExtra>
+                {isAgentsEnabled ? (
+                  /* The global FAB is intentionally hidden while a modal overlay
+                      is open, so traces expose a local PXI entrypoint here. */
+                  <Button
+                    size="S"
+                    variant="primary"
+                    leadingVisual={<PxiGlyph variant="resting" />}
+                    onPress={() => setIsOpen(true)}
+                  >
+                    Ask PXI
+                  </Button>
+                ) : null}
+                <ShareLinkButton
+                  preserveSearchParams
+                  buttonText="Share"
+                  tooltipText="Copy trace link to clipboard"
+                  successText="Trace link copied to clipboard"
                 />
-              </Suspense>
-            </DialogContent>
-          )}
-        </Dialog>
-      </Modal>
-    </ModalOverlay>
+              </DialogTitleExtra>
+            </DialogHeader>
+            <Suspense fallback={<Loading />}>
+              <TraceDetails
+                traceId={traceId as string}
+                projectId={projectId as string}
+              />
+            </Suspense>
+          </DialogContent>
+        )}
+      </Dialog>
+    </Drawer>
   );
 }
